Rename component, start_tables, home-utils, and soft-lock to snake_case
src/component -> src/ui_component, src/start_tables -> src/home_tables,
src/home-utils.js -> src/home_utils.js, src/soft-lock.js -> src/soft_lock.js.
Updates all imports, fetch paths, and comments that reference the old names.
