+++ src/ui.css
@@ -76,12 +76,12 @@
   --font-body:    -apple-system, BlinkMacSystemFont, "Segoe UI", Helvetica, Arial, sans-serif;
   --font-display: "Palatino Linotype", Palatino, "Book Antiqua", Georgia, serif;
 
-  /* Aliases consumed by <color-picker> (component/color-picker.js) */
+  /* Aliases consumed by <color-picker> (ui_component/color-picker.js) */
   --color-border-tertiary:  var(--border);
   --color-border-secondary: var(--border-2);
   --color-text-primary:     var(--text);
 
-  /* Aliases consumed by <range-ticked> (component/range_ticked.js) */
+  /* Aliases consumed by <range-ticked> (ui_component/range_ticked.js) */
   --color-accent:        var(--primary);
   --color-text-tertiary: var(--text-3);
 }
