+++ src/ui_component/color-picker.js
