+++ src/ui_component/example-color-picker.html
