+++ src/ui_component/example-range-ticked.html
