+++ src/ui_component/range_ticked.js
